requirements.txt

streamlit
tflite-runtime==2.11.0
Pillow
numpy
